+++ logger/config.go
@@ -11,7 +11,7 @@ import (
 func init() {
 	// logrus.SetReportCaller(true)
 
-	logrus.SetLevel(logrus.Level(viper.GetInt("LOGGER.LOG_LEVEL")))
+	logrus.SetLevel(logrus.Level(viper.GetInt64("LOGGER.LOG_LEVEL")))
 
 	f, err := os.OpenFile("iitkcoin.log", os.O_APPEND|os.O_CREATE|os.O_RDWR, 0666)
 	if err != nil {
