+++ util/otp.go
@@ -13,7 +13,7 @@ func RandomOTP() string {
 	return stringWithCharset(8, charset)
 }
 
-func stringWithCharset(length int, charset string) string {
+func stringWithCharset(length int64, charset string) string {
 	b := make([]byte, length)
 	for i := range b {
 		b[i] = charset[seededRand.Intn(len(charset))]
